requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv